Cambio de ubiacación de métodos HazOperación y Resolver pedido en la clases Protocolo
